requirements.txt
flask
numpy
mysql-connector-python
bs4
requests
itertools